Replace djangowind with django-cas-ng
Also, I updated the urls.py to the new path() syntax.

Some of the requirements were out of date here as well, there's been
some changes with boto and django-storages stuff.

I'm still working on a new version of ccnmtlsettings that will actually
have the new django-cas-ng settings. For now, my goal is to have
ccnmtlsettings support both django-cas-ng and djangowind. But ctldjango
should have just django-cas-ng, so we use that for any new apps we make.